+++ app/src/main/java/com/olamide/findartt/viewmodels/ArtworkViewModel.java
@@ -1,22 +1,109 @@
-//package com.olamide.findartt.viewmodels;
-//
-//import android.arch.lifecycle.LiveData;
-//import android.arch.lifecycle.ViewModel;
-//
-//import com.olamide.findartt.database.FindArttDatabase;
-//import com.olamide.findartt.models.Artwork;
-//
-//public class ArtworkViewModel extends ViewModel {
-//
-//    private LiveData<Artwork> artworkLiveData;
-//
-//
-//    public ArtworkViewModel(FindArttDatabase database, int artworkId) {
-//        artworkLiveData = database.artworkDao().loadArtworkById(artworkId);
-//    }
-//
-//
-//    public LiveData<Artwork> getArtworkLiveData() {
-//        return artworkLiveData;
-//    }
-//}
+package com.olamide.findartt.viewmodels;
+
+import android.app.Application;
+import androidx.lifecycle.AndroidViewModel;
+import androidx.lifecycle.MutableLiveData;
+import androidx.annotation.NonNull;
+
+import com.olamide.findartt.di.rx.SchedulersFactory;
+import com.olamide.findartt.models.Artwork;
+import com.olamide.findartt.models.mvvm.MVResponse;
+import com.olamide.findartt.utils.network.FindArttRepository;
+
+import io.reactivex.disposables.CompositeDisposable;
+
+public class ArtworkViewModel extends AndroidViewModel {
+
+    private FindArttRepository findArttRepository;
+    private SchedulersFactory schedulersFactory;
+
+    private final CompositeDisposable disposables = new CompositeDisposable();
+    private final CompositeDisposable favouriteDisposables = new CompositeDisposable();
+    private final CompositeDisposable videoDisposables = new CompositeDisposable();
+    private final MutableLiveData<MVResponse> artworkSummaryLive = new MutableLiveData<>();
+    private final MutableLiveData<MVResponse> artworkFavouriteLive = new MutableLiveData<>();
+
+
+
+    public ArtworkViewModel(Application application, FindArttRepository findArttRepository, SchedulersFactory schedulersFactory) {
+        this(application);
+        this.findArttRepository = findArttRepository;
+        this.schedulersFactory = schedulersFactory;
+    }
+
+    public ArtworkViewModel(@NonNull Application application) {
+        super(application);
+    }
+
+    public MutableLiveData<MVResponse> getArtWorkResponse() {
+        return artworkSummaryLive;
+    }
+
+    public MutableLiveData<MVResponse> getArtWorkFavourite() {
+        return artworkFavouriteLive;
+    }
+
+
+
+    public void findArtSummary(String accessToken, int artId ) {
+
+        disposables.add(findArttRepository.getArtSummary(accessToken, artId)
+                .subscribeOn(schedulersFactory.io())
+                .observeOn(schedulersFactory.ui())
+                .doOnSubscribe((d) -> artworkSummaryLive.setValue(MVResponse.loading()))
+                .subscribe(
+                        result -> artworkSummaryLive.setValue(MVResponse.success(result)),
+                        throwable -> artworkSummaryLive.setValue(MVResponse.error(throwable))
+                ));
+
+    }
+
+    public void findArtFavourite( int artId ) {
+
+        favouriteDisposables.add(findArttRepository.loadArtworkById( artId)
+                .subscribeOn(schedulersFactory.io())
+                .observeOn(schedulersFactory.ui())
+                .doOnSubscribe((d) -> artworkFavouriteLive.setValue(MVResponse.loading()))
+                .subscribe(
+                        result -> artworkFavouriteLive.setValue(MVResponse.success(result)),
+                        throwable -> artworkFavouriteLive.setValue(MVResponse.error(throwable))
+                ));
+
+    }
+
+    public void insertFavouriteArt(Artwork artwork) {
+
+
+        favouriteDisposables.add(findArttRepository.insertFavouriteArt(artwork)
+                .subscribeOn(schedulersFactory.io())
+                .observeOn(schedulersFactory.ui())
+                .doOnSubscribe((d) -> artworkFavouriteLive.setValue(MVResponse.loading()))
+                .subscribe(
+                        () -> artworkFavouriteLive.setValue(MVResponse.success(new Artwork())),
+                        throwable -> artworkFavouriteLive.setValue(MVResponse.error(throwable))
+                ));
+
+    }
+
+    public void deleteFavouriteArt( Artwork artwork ) {
+
+        favouriteDisposables.add(findArttRepository.deleteFavouriteArt(artwork)
+                .subscribeOn(schedulersFactory.io())
+                .observeOn(schedulersFactory.ui())
+                .doOnSubscribe((d) -> artworkFavouriteLive.setValue(MVResponse.loading()))
+                .subscribe(
+                        () -> artworkFavouriteLive.setValue(MVResponse.success(null)),
+                        throwable -> artworkFavouriteLive.setValue(MVResponse.error(throwable))
+                )
+        );
+
+    }
+
+
+    @Override
+    protected void onCleared() {
+        disposables.clear();
+        favouriteDisposables.clear();
+        videoDisposables.clear();
+    }
+}
