+++ app/src/main/java/com/olamide/findartt/viewmodels/DashboardViewModel.java
@@ -1,4 +1,4 @@
-package com.olamide.findartt;
+package com.olamide.findartt.viewmodels;
 
 import android.app.Application;
 import androidx.lifecycle.AndroidViewModel;
