+++ app/src/main/java/com/olamide/findartt/viewmodels/LoginViewModel.java
@@ -1,37 +1,41 @@
-package com.olamide.findartt;
+package com.olamide.findartt.viewmodels;
 
 import androidx.lifecycle.MutableLiveData;
 import androidx.lifecycle.ViewModel;
 
 import com.olamide.findartt.di.rx.SchedulersFactory;
 import com.olamide.findartt.models.TokenInfo;
-import com.olamide.findartt.models.UserSignup;
+import com.olamide.findartt.models.UserLogin;
 import com.olamide.findartt.models.mvvm.MVResponse;
 import com.olamide.findartt.utils.network.FindArttRepository;
 
 import io.reactivex.disposables.CompositeDisposable;
 
-public class SignUpViewModel extends ViewModel {
+
+
+public class LoginViewModel extends ViewModel {
 
     private FindArttRepository findArttRepository;
     private final CompositeDisposable disposables = new CompositeDisposable();
     private final MutableLiveData<MVResponse> responseLiveData = new MutableLiveData<>();
-    private SchedulersFactory schedulersFactory;
+    private  SchedulersFactory schedulersFactory;
 
 
-    public SignUpViewModel(FindArttRepository findArttRepository, SchedulersFactory schedulersFactory) {
+    public LoginViewModel(FindArttRepository findArttRepository, SchedulersFactory schedulersFactory) {
         this.findArttRepository = findArttRepository;
         this.schedulersFactory = schedulersFactory;
     }
 
-    public MutableLiveData<MVResponse> getSignupResponse() {
+    public MutableLiveData<MVResponse> getLoginResponse() {
         return responseLiveData;
     }
 
+    /*
+     * method to call normal login api with $(email + password)
+     * */
+    public void hitLogin(UserLogin userLogin) {
 
-    public void signUp(UserSignup userSignup) {
-
-        disposables.add(findArttRepository.signUp(userSignup)
+        disposables.add(findArttRepository.login(userLogin)
                 .subscribeOn(schedulersFactory.io())
                 .observeOn(schedulersFactory.ui())
                 .doOnSubscribe((d) -> responseLiveData.setValue(MVResponse.loading()))
@@ -42,9 +46,22 @@ public class SignUpViewModel extends ViewModel {
 
     }
 
-    public void signUpGoogle(TokenInfo tokenInfo) {
+    public void hitGoogleLogin(TokenInfo tokenInfo) {
 
-        disposables.add(findArttRepository.signUpGoogle(tokenInfo)
+        disposables.add(findArttRepository.loginGoogle(tokenInfo)
+                .subscribeOn(schedulersFactory.io())
+                .observeOn(schedulersFactory.ui())
+                .doOnSubscribe((d) -> responseLiveData.setValue(MVResponse.loading()))
+                .subscribe(
+                        result -> responseLiveData.setValue(MVResponse.success(result)),
+                        throwable -> responseLiveData.setValue(MVResponse.error(throwable))
+                ));
+
+    }
+
+    public void getUserFromToken(String accessToken) {
+
+        disposables.add(findArttRepository.getUserFromToken(accessToken)
                 .subscribeOn(schedulersFactory.io())
                 .observeOn(schedulersFactory.ui())
                 .doOnSubscribe((d) -> responseLiveData.setValue(MVResponse.loading()))
@@ -59,6 +76,5 @@ public class SignUpViewModel extends ViewModel {
     protected void onCleared() {
         disposables.clear();
     }
-
-
 }
+
